Consertado erro na tela de produto, travado digitar , e letras no campo peso, tela de usuario quase finalizada.

+++ ferragem/src/beans/Produto.java
@@ -12,15 +12,15 @@ package beans;
 public class Produto {
     private int idProduto;
     private String descricao;
-    private float peso;
+    private double peso;
      
 
     
-    public float getPeso() {
+    public double getPeso() {
         return peso;
     }
 
-    public void setPeso(float peso) {
+    public void setPeso(double peso) {
         this.peso = peso;
     }
